restructured database models
